backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
exifread==3.0.0
requests==2.31.0
transformers==4.35.0
dashscope==1.10.0  # 阿里云百炼/Qwen访问库
pydantic==2.4.2
python-dotenv==1.0.0